requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-property==0.2.2
async-timeout==5.0.1
attrs==24.2.0
audeer==2.2.0
audiofile==1.5.0
audioop-lts==0.2.1
audmath==1.4.1
base91==1.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
discord-pretty-help==2.0.7
discord.py==2.4.0
distro==1.9.0
ffmpeg==1.4
fonttools==4.55.3
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
instaloader==4.14
jiter==0.8.2
kiwisolver==1.4.7
matplotlib==3.10.0
mojang==1.1.0
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.3
openai==1.59.3
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
propcache==0.2.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
PyNaCl==1.5.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytubefix==8.12.0
pytz==2024.2
requests==2.32.3
scipy==1.14.1
scrapetube==2.5.1
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.16.0
youtube-dl==2021.12.17
youtube-dlc==2020.11.11.post3
yt-dlp==2024.12.23
